Add files via upload
This version lets users modify the color of the rectangles. This information is then retained in the CSV output (and will also be imported when one continues to work on a file).

In this version I also adjusted the sorting of exported columns; the first six columns are now always in this order: Index, Page, Coordinates, Color, Anchor, Source
Importing files that do not have this column structure will fail, which also pertains to CSV output created by previous versions of this program. However, one can create the appropriate column structure in excel or something similar. For the new Color column, one should add "red" in each cell. After that, the import should work again.